backend/requirements.txt
# FastAPI and server dependencies
fastapi
uvicorn[standard]
python-multipart
websockets

# Database dependencies
sqlalchemy
psycopg2-binary
asyncpg
alembic

# LangChain and AI dependencies
langchain
langchain-community
langchain-ollama
langchain-core

# Vector database
chromadb-client

# HTTP requests and utilities
requests
pydantic

# Development
httpx

# Redis dependency
redis


crawl4ai
nest_asyncio

# Graph-based verification dependencies
numpy<2.0.0
neo4j

# ML dependencies
torch-geometric
spacy

# Bayesian and statistical analysis
scipy
pymc
arviz[all]
scikit-learn

# Time series and temporal analysis
python-dateutil
